Move `CacheConfig` from `config/__init__.py` to `config/cache.py`
